Tableau Data Structure
Coded most of the operations/functionality I want for the tableau data
type so far such as pivoting, redoing/undoing, etc. May change the
structure soon to having Tableau as an interface and having
TuckerTableau and SimplexTableau as data types that implement Tableau
(then they'd need constructors that accept the opposite type to allow
the conversion instead).

Also did a tiny bit on starting to setup a GUI, but not much, and not
final at all.